fix compile errors from List<Object?> casts and extract StatelessWidgets for avoid-returning-widgets
- Fix for-in type errors: add .cast<int/String/Map>() on List<Object?> fields
  (SpecialRule.list, monster/monster_instance conditions, scenario sections)
- Fix Object? assignment errors in room.dart, game_save_state.dart, scenario_methods.dart
- Replace widget-returning methods in set_level_menu.dart with proper StatelessWidget
  classes (_LevelButton, _DifficultyButton, _LevelLegend)

+++ frosthaven_assistant/lib/Resource/commands/next_round_command.dart
@@ -51,7 +51,7 @@ class NextRoundCommand extends Command {
 
     for (var rule in _gameState.scenarioSpecialRules) {
       if (rule.type == "Timer" && !rule.startOfRound) {
-        for (int round in rule.list) {
+        for (int round in rule.list.cast<int>()) {
           //minus 1 means always
           if (round == _gameState.round.value || round == -1) {
             if (_settings.showReminders.value) {
@@ -67,7 +67,7 @@ class NextRoundCommand extends Command {
     //start of next round is now
     for (var rule in _gameState.scenarioSpecialRules) {
       if (rule.type == "Timer" && rule.startOfRound) {
-        for (int round in rule.list) {
+        for (int round in rule.list.cast<int>()) {
           //minus 1 means always
           final toastMessage = _gameState.toastMessage.value;
           if (round - 1 == _gameState.round.value || round == -1) {
