+++ frosthaven_assistant/lib/Resource/state/monster_instance.dart
@@ -169,7 +169,7 @@ class MonsterInstance extends FigureState {
 
     if (json.containsKey("conditionsAddedThisTurn")) {
       List<Object?> condis2 = json["conditionsAddedThisTurn"] as List<Object?>;
-      for (int item in condis2) {
+      for (int item in condis2.cast<int>()) {
         if (item >= 0 && item < Condition.values.length) {
           _conditionsAddedThisTurn.add(Condition.values[item]);
         }
@@ -177,7 +177,7 @@ class MonsterInstance extends FigureState {
     }
     if (json.containsKey("conditionsAddedPreviousTurn")) {
       List<Object?> condis3 = json["conditionsAddedPreviousTurn"] as List<Object?>;
-      for (int item in condis3) {
+      for (int item in condis3.cast<int>()) {
         if (item >= 0 && item < Condition.values.length) {
           _conditionsAddedPreviousTurn.add(Condition.values[item]);
         }
